+++ WebAPI/Services/UserService.cs
@@ -7,14 +7,18 @@ namespace WebAPI.Services
     public class UserService : IUserService
     {
         IUserRepository userRepository;
+        IRentalService rentalService;
 
-        public UserService(IUserRepository userRepository)
+        public UserService(IUserRepository userRepository, IRentalService rentalService)
         {
             this.userRepository = userRepository;
+            this.rentalService = rentalService;
         }
 
         public User Delete(string Id, CancellationToken cancellationToken)
         {
+            if (rentalService.HasRelationship(GetById(Id, cancellationToken)))
+                return null;
             return userRepository.Delete(Id);
         }
